requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
flask
flask-ngrok
Pillow
gunicorn
